==> 02_BleachingMetrics.Rmd <==
---
title: "Calculation of Physiological Bleaching Metrics"
author: "Serena Hackerott"
date: "7/19/2024"
output:
  html_document:
    toc: yes
    df_print: paged
  html_notebook:
    toc: yes
    toc_float: yes
---

# Setup

```{r Setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```


### Load Packages
```{r}
##Install Packages if Needed
if (!require("ggplot2")) install.packages("ggplot2")

##Load Packages
library(ggplot2)
```
Note: Run "Graphing Parameters" section from 01_ExperimentalSetup.R file

# Sample Data

### Load and Organize Data
```{r}
####Load Data####
Sym<-read.csv("Data/Symbionts.csv", header=TRUE)
Chl<-read.csv("Data/Chlorophyll.csv", header=TRUE)
Wax.Stand<-read.csv("Data/WaxStandards.csv", header=TRUE)
SampData<-read.csv("Data/SampleData.csv", header=TRUE)

##Set factor variables
SampData$TimeP<-factor(SampData$TimeP, levels=c("W2", "M1", "M4"), ordered=TRUE)
SampData$Site<-factor(SampData$Site, levels=c("KL", "SS"), ordered=TRUE)
SampData$Genotype<-factor(SampData$Genotype, levels=c("AC8", "AC10", "AC12"), ordered=TRUE)
SampData$Treatment<-factor(SampData$Treatment, levels=c("Control", "Heat"), ordered=TRUE)
SampData$Treat<-factor(SampData$Treat, levels=c("C", "H"), ordered=TRUE)

##Add a Sample Set Variable
SampData$Set<-paste(SampData$TimeP, SampData$Site, SampData$Genotype, SampData$Treat, sep=".")

##Add a Season Variable
SampData$Seas<-ifelse(SampData$TimeP=="W2", "Summer1", ifelse(SampData$TimeP=="M1", "Summer2", ifelse(SampData$TimeP=="M4", "Winter" , NA)))
SampData$Seas<-factor(SampData$Seas, levels=c("Summer1", "Summer2", "Winter"), ordered=TRUE)

```


### Sample Surface Area
```{r}
####Create Standard Curve

##Calculate Surface Area of Cylinders from Standards
Wax.Stand$SA_cm2<-2*pi*(Wax.Stand$Dm_cm/2)*Wax.Stand$Ht_cm+2*pi*(Wax.Stand$Dm_cm/2)^2

##Calculate Difference in Wax Weight
Wax.Stand$Wax.D_g<- Wax.Stand$Wax.F_g-Wax.Stand$Wax.I_g

##Linear Model of SA as a Function of Wax Weight
wax.SA.lm  <- lm(SA_cm2~Wax.D_g, data=Wax.Stand)
summary(wax.SA.lm)
coef(wax.SA.lm)

SA.mod <- function(wax.weight) {
  coefs <- coef(wax.SA.lm)
  #y = mx + b
  SA <- ((coefs[2] * wax.weight) + coefs[1])
  return(SA)}

####Apply Standard Curve

##Calculate Difference in Wax Weight (g) in Sample Data
SampData$Wax.D_g<- SampData$Wax.F_g-SampData$Wax.I_g

##Calculate Surface Area (cm^2) by Applying Linear Model Function
SampData$SA_cm2<-SA.mod(SampData$Wax.D_g)

```


# Bleaching Metrics

### Chlorophyll

#### Calculate Chlorophyll Concentration
Equations for Dinos from Jeffrey and Humphrey 1975 in 100% acetone
Chla = 11.43 * A663 - 0.64 * A630
Chlc2 = 27.09 * A630 - 3.63 * A663
```{r}
##Subtract Background A750 from A630 and A663
Chl$A630.c<-Chl$A630-Chl$A750
Chl$A663.c<-Chl$A663-Chl$A750

##Divide by Pathlength (0.5cm pathlength for 175ul sample in UVStar Plate) 
Chl$A630.c<-c(Chl$A630.c/0.5)
Chl$A663.c<-c(Chl$A663.c/0.5)

##Calculate Chl-a and Chl-c2 in ug/ml
Chl$Chl.a_ug.ml<-11.43*Chl$A663.c- 0.64*Chl$A630.c
Chl$Chl.c2_ug.ml <- 27.09*Chl$A630.c - 3.63*Chl$A663.c

##Merge with Sample Data to Calculate Chlorophyll per Surface Area
#Merges by Random Number (RandN) column
#Adds necessary Slurry Volume (Vol_ml) and Surface Area (SA_cm2) columns
#Retains only Color Score project samples
Chl<-merge(Chl, SampData,  all.x=FALSE, all.y=TRUE)

##Calculate Total Chlorophyll-a and c2 (ug) 
Chl$Chl.a_ug<-Chl$Chl.a_ug.ml*Chl$Vol_ml
Chl$Chl.c2_ug<-Chl$Chl.c2_ug.ml*Chl$Vol_ml

##Calculate Chlorophyll-a and c2 per Surface Area (ug/cm^2)
Chl$Chl.a_ug.cm2<-Chl$Chl.a_ug/Chl$SA_cm2
Chl$Chl.c2_ug.cm2<-Chl$Chl.c2_ug/Chl$SA_cm2

##Initial Visual Check
ggplot(Chl, aes(x=Set, y=Chl.a_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

ggplot(Chl, aes(x=Set, y=Chl.c2_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

##Set Negative Values to Zero
Chl$Chl.a_ug.cm2[(which(Chl$Chl.a_ug.cm2<0))]<-0
Chl$Chl.c2_ug.cm2[(which(Chl$Chl.c2_ug.cm2<0))]<-0

##Add Chlorophyll-a and c2 for Total Chlorophyll per Surface Area (ug/cm^2)
Chl$Chl_ug.cm2<-Chl$Chl.a_ug.cm2+Chl$Chl.c2_ug.cm2

##Visual Check
ggplot(Chl, aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

```

#### Check for Outliers
Check for Outliers by Treatment and Timepoint, grouped by sample set. Removing clear outlier samples and outliers of technical replicates before averaging across technical repliates. 

Control Treatment
```{r}
##Subset Control Treatment in Thermal Tolerance Assay 
Chl.C<-subset(Chl, Treat=="C")

#W2
ggplot(subset(Chl.C, TimeP=="W2"), aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,2.5)+
  theme(axis.text.x = element_text(angle = 90))

Chl.C$RandN[which(Chl.C$TimeP=="W2" & Chl.C$Chl_ug.cm2<0.25)] #"W2_45"

#M1
ggplot(subset(Chl.C, TimeP=="M1"), aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,2.5)+
  theme(axis.text.x = element_text(angle = 90))

#M4
ggplot(subset(Chl.C, TimeP=="M4"), aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,2.5)+
  theme(axis.text.x = element_text(angle = 90))


##Remove Outliers 
Chl.C.o<-Chl.C[-c(which(Chl.C$TimeP=="W2" & Chl.C$Chl_ug.cm2<0.25)),]

```


Heated Treatment
```{r}
##Subset Heated Treatment in Thermal Tolerance Assay
Chl.H<-subset(Chl, Treat=="H")

#W2
ggplot(subset(Chl.H, TimeP=="W2"), aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,1)+
  theme(axis.text.x = element_text(angle = 90))

Chl.H$RandN[which(Chl.H$TimeP=="W2" & Chl.H$Chl_ug.cm2>0.75)] #"W2_88" "W2_88" "W2_88"

#M1
ggplot(subset(Chl.H, TimeP=="M1"), aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,1)+
  theme(axis.text.x = element_text(angle = 90))

#M4
ggplot(subset(Chl.H, TimeP=="M4"), aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,1)+
  theme(axis.text.x = element_text(angle = 90))

Chl.H$RandN[which(Chl.H$TimeP=="M4" & Chl.H$Chl_ug.cm2<.25 & Chl.H$Genotype=="AC12")] #"M4_72"
Chl.H$RandN[which(Chl.H$TimeP=="M4" & Chl.H$Chl_ug.cm2>.55 & Chl.H$Site=="KL")] #"M4_72"
Chl.H$RandN[which(Chl.H$TimeP=="M4" & Chl.H$Chl_ug.cm2>.5 & Chl.H$Genotype=="AC10")] #"M4_54"


##Remove Outliers
Chl.H.o<-Chl.H[-c(which((Chl.H$TimeP=="W2" & Chl.H$Chl_ug.cm2>0.75) |
                          (Chl.H$TimeP=="M4" & Chl.H$Chl_ug.cm2<.25 & Chl.H$Genotype=="AC12") |
                          (Chl.H$TimeP=="M4" & Chl.H$Chl_ug.cm2>.55 & Chl.H$Site=="KL") |
                          (Chl.H$TimeP=="M4" & Chl.H$Chl_ug.cm2>.5 & Chl.H$Genotype=="AC10"))),]


```


#### Add Chlorophyll to Bleaching Data
```{r}
##Recombine Cleaned Chlorophyll dataframes
Chl.o<-rbind(Chl.C.o, Chl.H.o)

ggplot(Chl.o, aes(x=Set, y=Chl_ug.cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))


##Average Across Replicate Readings ("Rep" column: A, B, C)
names(Chl.o)
Chl.a<-aggregate(Chl.o$Chl_ug.cm2, list(Chl.o$RandN, Chl.o$ID), mean)
names(Chl.a)<-c("RandN", "ID", "Chl_ug.cm2")


##Add Total Chlorophyll to Bleaching Metric Data

##Merge Averaged Chlorophyll Data with Sample Meta Data to create Bleaching Data dataframe
#Merges by Random Number (RandN) and ID columns
#Retains all Thermal Tolerance Assay samples
#Retains RandN, ID, TimeP, Site, Genotype, Treat, Treatment, Set, Season, and SA_cm2 from SampData
names(SampData)
BleachData<-merge(SampData[,c(1:7, 12, 11, 14)], Chl.a, all.x=TRUE, all.y=FALSE)

```



### Symbionts

#### Calculate Symbiont Density
```{r}
####Calculate Dilution Factors

#Scales for 1:10 Dilution Sent for Flow Cytometry
Sym$Send_df<-Sym$SendInputVol_ul/Sym$SendTotalVol_ul

##Scales for Resuspending Symbiont Pellet based on Volume of 0.01% SDS Used
Sym$Resp_df<-Sym$InputVol_ul/Sym$RespVol_ul

####Calculate Symbiont Density

##Scale Flow Cy Count/ul to account for Dilution Factor applied at Flow Cy Center
Sym$Count_ul_scaled_FC_df<-Sym$Count_ul/Sym$FC_df

##Scale Flow Cy Count/ul to account for Sent Dilution Factor
Sym$Count_ul_scaled_send_df<-Sym$Count_ul_scaled_FC_df/Sym$Send_df

##Scale Flow Cy Count/ul to account for Resuspension Dilution Factor
Sym$Count_ul_scaled_resp_df<-Sym$Count_ul_scaled_send_df/Sym$Resp_df

##Merge with Sample Data to Calculate Symbionts per Surface Area
#Merges by Random Number (RandN) column
#Adds necessary Slurry Volume (Vol_ml) and Surface Area (SA_cm2) columns
#Retains only Color Score project samples
Sym<-merge(Sym, SampData, all.x=FALSE, all.y=TRUE)

##Calculate Total Symbiont Count (scaling to Slurry Total Volume in ul)
Sym$Count_total<-Sym$Count_ul_scaled_resp_df*(Sym$Vol_ml*1000)

##Calculate Symbionts per Surface Area
Sym$Sym_cm2<-Sym$Count_total/Sym$SA_cm2

##Scale to Symb * 10^6 / cm^2
Sym$Sym10.6_cm2<-Sym$Sym_cm2/10^6

##Initial Visual Check
ggplot(Sym, aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

```


#### Check for Outliers
Check for Outliers by Treatment and Timepoint, grouped by sample set. Removing outliers of technical replicates before averaging across technical repliates. 

Control Treatment
```{r}
##Subset Control Treatment in Thermal Tolerance Assay
Sym.C<-subset(Sym, Treat=="C")

#W2
ggplot(subset(Sym.C, TimeP=="W2"), aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,2)+
  theme(axis.text.x = element_text(angle = 90))

#M1
ggplot(subset(Sym.C, TimeP=="M1"), aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,2)+
  theme(axis.text.x = element_text(angle = 90))

#M4
ggplot(subset(Sym.C, TimeP=="M4"), aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,2)+
  theme(axis.text.x = element_text(angle = 90))

```


Heated Treatment
```{r}
##Subset Heated Treatment in Thermal Tolerance Assay
Sym.H<-subset(Sym, Treat=="H")

#W2
ggplot(subset(Sym.H, TimeP=="W2"), aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,1)+
  theme(axis.text.x = element_text(angle = 90))

#M1
ggplot(subset(Sym.H, TimeP=="M1"), aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,1)+
  theme(axis.text.x = element_text(angle = 90))

Sym.H$RandN[which(Sym.H$TimeP=="M1" & Sym.H$Sym10.6_cm2>0.75)] #"M1_73" "M1_73"

#M4
ggplot(subset(Sym.H, TimeP=="M4"), aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  ylim(0,1.5)+
  theme(axis.text.x = element_text(angle = 90))

Sym.H$RandN[which(Sym.H$TimeP=="M4" & Sym.H$Sym10.6_cm2>0.75)] #"M4_61" 


##Remove Outliers
Sym.H.o<-Sym.H[-c(which((Sym.H$TimeP=="M1" & Sym.H$Sym10.6_cm2>0.75) | 
                          (Sym.H$TimeP=="M4" & Sym.H$Sym10.6_cm2>0.75))),]

```


#### Add Symbionts to Bleaching Data
```{r}
##Recombine Cleaned Chlorophyll dataframes
Sym.o<-rbind(Sym.C, Sym.H.o)

ggplot(Sym.o, aes(x=Set, y=Sym10.6_cm2)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))


####Average Across Replicate Readings ("Rep" column: A, B, C)
names(Sym.o)
Sym.a<-aggregate(Sym.o$Sym10.6_cm2, list(Sym.o$RandN, Sym.o$ID), mean)
names(Sym.a)<-c("RandN", "ID", "Sym10.6_cm2")

####Add Symbiont Density to Bleaching Metric Data

##Merge Averaged Symbiont Data with Bleaching Metric Data
#Merges by Random Number (RandN) and ID columns
#Retains all Thermal Tolerance Assay samples
names(BleachData)
BleachData<-merge(BleachData, Sym.a, all.x=TRUE, all.y=FALSE)

```


### Write Out Bleaching Data
```{r}
write.csv(BleachData, "Outputs/BleachData.csv", row.names=FALSE)

```
