==> 03_ColorScore.Rmd <==
---
title: "Calculation of Color Score"
author: "Serena Hackerott and Lauren Gregory"
date: "7/19/2024"
output:
  html_document:
    toc: yes
    df_print: paged
  html_notebook:
    toc: yes
    toc_float: yes
---

# Setup
```{r Setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```


### Load Packages
```{r}
##Install Packages if Needed
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("vegan")) install.packages("vegan")
if (!require("FactoMineR")) install.packages("FactoMineR")
if (!require("factoextra")) install.packages("factoextra")
if (!require("PerformanceAnalytics")) install.packages("PerformanceAnalytics")
if (!require("cowplot")) install.packages("cowplot")
if (!require("png")) install.packages("png")
if (!require("grid")) install.packages("grid")
if (!require("MASS")) install.packages("MASS")

##Load Packages
library("ggplot2")
library("vegan")
library("FactoMineR")
library("factoextra")
library("PerformanceAnalytics")
library("cowplot")
library("png")
library("grid")
library("MASS")


```
Note: Run "Graphing Parameters" section from 01_ExperimentalSetup.R file


### Load and Organize Data
```{r}
##Load Data
Color<-read.csv("Data/Color.csv", header=TRUE)
SampData<-read.csv("Data/SampleData.csv", header=TRUE)

##Set factor variables
SampData$TimeP<-factor(SampData$TimeP, levels=c("W2", "M1", "M4"), ordered=TRUE)
SampData$Site<-factor(SampData$Site, levels=c("KL", "SS"), ordered=TRUE)
SampData$Genotype<-factor(SampData$Genotype, levels=c("AC8", "AC10", "AC12"), ordered=TRUE)
SampData$Treatment<-factor(SampData$Treatment, levels=c("Control", "Heat"), ordered=TRUE)
SampData$Treat<-factor(SampData$Treat, levels=c("C", "H"), ordered=TRUE)

##Add a Sample Set Variable
SampData$Set<-paste(SampData$TimeP, SampData$Site, SampData$Genotype, SampData$Treat, sep=".")

##Add a Season Variable
SampData$Seas<-ifelse(SampData$TimeP=="W2", "Summer1", ifelse(SampData$TimeP=="M1", "Summer2", ifelse(SampData$TimeP=="M4", "Winter" , NA)))
SampData$Seas<-factor(SampData$Seas, levels=c("Summer1", "Summer2", "Winter"), ordered=TRUE)

##Set rownames to ID
rownames(SampData)<-SampData$ID
rownames(Color)<-Color$ID

##Merge Color data with Sample Meta Data
Color<-merge(Color, SampData, all.x=FALSE, all.y=FALSE)
rownames(Color)<-Color$ID

```


# Calculate Color Score

### Standardize Colors
Standardize RGB colors by dividing Coral color by color standards. 
```{r}
Color$Red.Norm.Coral <- Color$Red.Coral/Color$Red.Standard
Color$Green.Norm.Coral <- Color$Green.Coral/Color$Green.Standard
Color$Blue.Norm.Coral <- Color$Blue.Coral/Color$Blue.Standard

```

# Color Score Full Dataset

### Run PCA
```{r}
Color.PCA<-prcomp(Color[,c(19:21)], scale.=TRUE) 

#Initial plot
fviz_pca_ind(Color.PCA)

#Check Variance Explained by Components
summary(Color.PCA)

#Visualize the importance of each principal component
fviz_eig(Color.PCA, addlabels = TRUE) 

```

PC1 Explains 94.9% of the variance in the color data.


#### Axes and Scores
```{r}
##Extract % Variance PC 1
PCA.PC1<-sprintf("%1.2f",get_eigenvalue(Color.PCA)$variance.percent[1])
PCA.PC1

##Extract % Variance PC 2
PCA.PC2<-sprintf("%1.2f",get_eigenvalue(Color.PCA)$variance.percent[2])
PCA.PC2

##Extract Individual Sample Scores
Color.PCA.scores <- as.data.frame(Color.PCA$x)
Color.PCA.scores$ID<-rownames(Color.PCA.scores)

#Prepare for Plotting
Color.PCA.scores<-merge(Color.PCA.scores, SampData)

```


#### Plot PCA 
```{r}
#Plot PCA
Color.PCA.plot<-ggplot(data = Color.PCA.scores, aes(x = PC1, y = PC2)) + 
  geom_point(aes(colour = Treatment), size = point.sz-1, alpha = 0.8) + 
  scale_colour_manual(values =HC.colors.o)+
  theme_classic()+
  scale_x_continuous(limits = c(-3, 5.5))+
  scale_y_continuous(limits = c(-4.25, 4.25))+
  labs(x=paste0('PC 1 (',PCA.PC1,"%)"), y=paste0('PC 2 (',PCA.PC2,"%)"))+
  ggtitle("Color Score PCA")+
   theme(plot.title=element_text(size=plot.title.sz, hjust=0.5),
     axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz),
        legend.title=element_text(size=leg.title.sz),
        legend.box.background = element_rect(color = "black"), 
        legend.position="bottom");Color.PCA.plot

```

### LDA on PCA Scores
```{r}
##LDA on PCA Scores
Color.LDA_PCA<-lda(Treatment~PC1+PC2+PC3, data=Color.PCA.scores)

##Predict
Color.pLDA_PCA<-predict(object=Color.LDA_PCA, newdata=Color.PCA.scores)

##Save LD1 Scores
Color.LDA_PCA.scores<-data.frame(ID=Color.PCA.scores$ID, LD1=Color.pLDA_PCA$x)

#Prepare for Plotting
Color.LDA_PCA.scores<-merge(Color.LDA_PCA.scores, SampData)

```


#### LD 1 Density Plot

```{r}
Color.LDA.PCA_LD1.plot<-ggplot(data = Color.LDA_PCA.scores, aes(x = LD1)) + 
  geom_density(aes(fill = Treatment), alpha = 0.6) + 
  scale_fill_manual(values =HC.colors.o)+
  theme_classic()+
  scale_x_continuous(limits = c(-3, 5.5))+
  labs(x="LD1", y= "Density")+
  ggtitle("Full Dataset")+
   theme(plot.title=element_text(size=plot.title.sz, hjust=0.5),
     axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz),
        legend.title=element_text(size=leg.title.sz),
        legend.box.background = element_rect(color = "black"), 
        legend.position="bottom");Color.LDA.PCA_LD1.plot
```


### Extract Color Score
```{r}
##Retain Metadata columns of interest
names(Color.LDA_PCA.scores)
ColorData<-Color.LDA_PCA.scores[,c("ID", "TimeP", "Site", "Genotype", "Treat", "Treatment", "Seas", "Set")] 

##Retain PC1 as Color Score
ColorData$Score_Full<-Color.LDA_PCA.scores$LD1

##Initial Visual Check
ggplot(ColorData, aes(x=Set, y=Score_Full)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

##Invert signs for Control > Heated
ColorData$Score_Full<-ColorData$Score_Full*(-1)

#Adding 20 to make all score values positive 
ColorData$Score_Full<- ColorData$Score_Full +20

##Initial Visual Check
ggplot(ColorData, aes(x=Set, y=Score_Full)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

##Plot by Treatment
ggplot(ColorData, aes(x=Treatment, y=Score_Full)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

```


# Color Score By Timepoint
Repeat the calculation of Color Score for each Timepoint individually

### Summer 1 Aug 2021 W2

#### Run PCA
```{r}
#Subset W2 Timepoint
Color_W2<-subset(Color, TimeP=="W2")

#PCA
Color_W2.PCA<-prcomp(Color_W2[,c(19:21)], scale.=TRUE) 

#Initial plot
fviz_pca_ind(Color_W2.PCA)

#Check Variance Explained by Components
summary(Color_W2.PCA)

#Visualize the importance of each principal component
fviz_eig(Color_W2.PCA, addlabels = TRUE) 

```

PC1 Explains 97.9% of the variance in the color data.


### LDA on PCA Scores
```{r}
##Extract Individual Sample Scores
Color_W2.PCA.scores <- as.data.frame(Color_W2.PCA$x)
Color_W2.PCA.scores$ID<-rownames(Color_W2.PCA.scores)

##Add Treatment Variables
Color_W2.PCA.scores<-merge(Color_W2.PCA.scores, SampData, all.y=FALSE)

##LDA on PCA Scores
Color_W2.LDA_PCA<-lda(Treatment~PC1+PC2+PC3, data=Color_W2.PCA.scores)

##Predict
Color_W2.pLDA_PCA<-predict(object=Color_W2.LDA_PCA, newdata=Color_W2.PCA.scores)

##Save LD1 Scores
Color_W2.LDA_PCA.scores<-data.frame(ID=Color_W2.PCA.scores$ID, LD1=Color_W2.pLDA_PCA$x)

#Prepare for Plotting
Color_W2.LDA_PCA.scores<-merge(Color_W2.LDA_PCA.scores, SampData)

```


#### LD 1 Density Plot

```{r}
Color_W2.LDA.PCA_LD1.plot<-ggplot(data = Color_W2.LDA_PCA.scores, aes(x = LD1)) + 
  geom_density(aes(fill = Treatment), alpha = 0.6) + 
  scale_fill_manual(values =HC.colors.o)+
  theme_classic()+
  scale_x_continuous(limits = c(-3, 5.5))+
  labs(x="LD1", y= "Density")+
  ggtitle("Summer 1")+
   theme(plot.title=element_text(size=plot.title.sz, hjust=0.5),
     axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz),
        legend.title=element_text(size=leg.title.sz),
        legend.box.background = element_rect(color = "black"), 
        legend.position="bottom");Color_W2.LDA.PCA_LD1.plot
```

Samples align along LD1 according to Treatment (Heated vs Control). Coordinates along LD1 will be extracted as the Color Score. Coordinates will multiplied by -1 for intuitive interpretation (Control > Heated) then made positive by adding 20


### Summer 2 Aug 2021 M1

#### Run PCA
```{r}
#Subset W2 Timepoint
Color_M1<-subset(Color, TimeP=="M1")

#PCA
Color_M1.PCA<-prcomp(Color_M1[,c(19:21)], scale.=TRUE) 

#Initial plot
fviz_pca_ind(Color_M1.PCA)

#Check Variance Explained by Components
summary(Color_M1.PCA)

#Visualize the importance of each principal component
fviz_eig(Color_M1.PCA, addlabels = TRUE) 

```

PC1 Explains 96.2% of the variance in the color data.


### LDA on PCA Scores
```{r}
##Extract Individual Sample Scores
Color_M1.PCA.scores <- as.data.frame(Color_M1.PCA$x)
Color_M1.PCA.scores$ID<-rownames(Color_M1.PCA.scores)

##Add Treatment Variables
Color_M1.PCA.scores<-merge(Color_M1.PCA.scores, SampData, all.y=FALSE)

##LDA on PCA Scores
Color_M1.LDA_PCA<-lda(Treatment~PC1+PC2+PC3, data=Color_M1.PCA.scores)

##Predict
Color_M1.pLDA_PCA<-predict(object=Color_M1.LDA_PCA, newdata=Color_M1.PCA.scores)

##Save LD1 Scores
Color_M1.LDA_PCA.scores<-data.frame(ID=Color_M1.PCA.scores$ID, LD1=Color_M1.pLDA_PCA$x)

#Prepare for Plotting
Color_M1.LDA_PCA.scores<-merge(Color_M1.LDA_PCA.scores, SampData)

```


#### LD 1 Density Plot

```{r}
Color_M1.LDA.PCA_LD1.plot<-ggplot(data = Color_M1.LDA_PCA.scores, aes(x = LD1)) + 
  geom_density(aes(fill = Treatment), alpha = 0.6) + 
  scale_fill_manual(values =HC.colors.o)+
  theme_classic()+
  scale_x_continuous(limits = c(-3.5, 5))+
  labs(x="LD1", y= "Density")+
  ggtitle("Summer 2")+
   theme(plot.title=element_text(size=plot.title.sz, hjust=0.5),
     axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz),
        legend.title=element_text(size=leg.title.sz),
        legend.box.background = element_rect(color = "black"), 
        legend.position="bottom");Color_M1.LDA.PCA_LD1.plot
```

Samples align along LD1 according to Treatment (Heated vs Control). Coordinates along LD1 will be extracted as the Color Score. Coordinates will multiplied by -1 for intuitive interpretation (Control > Heated) then made positive by adding 20


### Winter Dec 20201 M4

#### Run PCA
```{r}
#Subset W2 Timepoint
Color_M4<-subset(Color, TimeP=="M4")

#PCA
Color_M4.PCA<-prcomp(Color_M4[,c(19:21)], scale.=TRUE) 

#Initial plot
fviz_pca_ind(Color_M4.PCA)

#Check Variance Explained by Components
summary(Color_M4.PCA)

#Visualize the importance of each principal component
fviz_eig(Color_M4.PCA, addlabels = TRUE) 

```

PC1 Explains 94.3% of the variance in the color data.


### LDA on PCA Scores
```{r}
##Extract Individual Sample Scores
Color_M4.PCA.scores <- as.data.frame(Color_M4.PCA$x)
Color_M4.PCA.scores$ID<-rownames(Color_M4.PCA.scores)

##Add Treatment Variables
Color_M4.PCA.scores<-merge(Color_M4.PCA.scores, SampData, all.y=FALSE)

##LDA on PCA Scores
Color_M4.LDA_PCA<-lda(Treatment~PC1+PC2+PC3, data=Color_M4.PCA.scores)

##Predict
Color_M4.pLDA_PCA<-predict(object=Color_M4.LDA_PCA, newdata=Color_M4.PCA.scores)

##Save LD1 Scores
Color_M4.LDA_PCA.scores<-data.frame(ID=Color_M4.PCA.scores$ID, LD1=Color_M4.pLDA_PCA$x)

#Prepare for Plotting
Color_M4.LDA_PCA.scores<-merge(Color_M4.LDA_PCA.scores, SampData)

```


#### LD 1 Density Plot

```{r}
Color_M4.LDA.PCA_LD1.plot<-ggplot(data = Color_M4.LDA_PCA.scores, aes(x = LD1)) + 
  geom_density(aes(fill = Treatment), alpha = 0.6) + 
  scale_fill_manual(values =HC.colors.o)+
  theme_classic()+
  scale_x_continuous(limits = c(-3.5, 5))+
  labs(x="LD1", y= "Density")+
  ggtitle("Winter")+
   theme(plot.title=element_text(size=plot.title.sz, hjust=0.5),
     axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        legend.text=element_text(size=leg.txt.sz),
        legend.title=element_text(size=leg.title.sz),
        legend.box.background = element_rect(color = "black"), 
        legend.position="bottom");Color_M4.LDA.PCA_LD1.plot
```

Samples align along LD1 according to Treatment (Heated vs Control). Coordinates along LD1 will be extracted as the Color Score. Coordinates will multiplied by -1 for intuitive interpretation (Control > Heated) then made positive by adding 20



### Extract Color Scores
```{r}
##Combine results from individual timepoints
ColorData.TP<-rbind(Color_W2.LDA_PCA.scores, Color_M1.LDA_PCA.scores, Color_M4.LDA_PCA.scores)

##Retain PC1 as Color Score
ColorData.TP$Score_TP<-ColorData.TP$LD1

##Invert signs for Control > Heated
ColorData.TP$Score_TP<-ColorData.TP$Score_TP*(-1)

#Adding 20 to make all score values positive 
ColorData.TP$Score_TP<- ColorData.TP$Score_TP +20

##Initial Visual Check
ggplot(ColorData.TP, aes(x=Set, y=Score_TP)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

##Plot by Treatment
ggplot(ColorData.TP, aes(x=Treatment, y=Score_TP)) + 
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA)+
  geom_jitter(shape=16, position=position_jitter(0.1))+
  theme(axis.text.x = element_text(angle = 90))

##Merge with Color Data
names(ColorData.TP)
ColorData<-merge(ColorData, ColorData.TP[,c("ID", "Score_TP")])
```


#### Correlation
```{r}
cor.test(ColorData$Score_Full, ColorData$Score_TP, method="spearman")
plot(ColorData$Score_Full, ColorData$Score_TP)
```

Calculations of Color score between the Full Dataset or Individual PCA's by Timepoints are significantly correlated with each other. 


### Write Out Color Score Data
```{r}
##Write out
write.csv(ColorData, "Outputs/ScoreData.csv", row.names=FALSE)

```


# Figures

### Figure 1 Panels d and e

#### PCA Inset
```{r}
##Update legend
Color.PCA.plot<-Color.PCA.plot+theme(legend.position=c(.5, .9), legend.direction="horizontal")

##Save PCA for Inset
ggsave(filename="Figures/Figure1d_inset_Color_Score_PCA.png", plot=Color.PCA.plot, dpi=300, width=5, height=5, units="in")
```


#### LDA with Inset
```{r}
##Update legend and title
Color.LDA.PCA_LD1.plot_fig1<-Color.LDA.PCA_LD1.plot+theme(legend.position="none")+ggtitle("Color Score PCA-LDA")

##Read in Priming Treatment Plot
PCA_Inset<-readPNG("Figures/Figure1d_inset_Color_Score_PCA.png")

####Plot Experimental Phases and Add Inset
Color.LDA.PCA_LD1.plot_fig1<-Color.LDA.PCA_LD1.plot_fig1 +
  ylim(-0.1, 0.6)+
  annotation_custom(rasterGrob(PCA_Inset), xmin=1.6, xmax=5.75, ymin=.22, ymax=.62)+
   annotate("segment", x = -3, xend = 5.5, y = -0.03, yend = -0.03, linewidth=bar.sz,
           arrow = arrow(ends = "both", angle = 90, length = unit(.2,"cm")))+
  annotate("text", x=1, y=-0.08, label="LD 1 Scores", size=levels.sz); Color.LDA.PCA_LD1.plot_fig1

```

#### Boxplot of Color Scores
```{r}
Score.boxplot_fig1<-ggplot(ColorData, aes(x=Treatment, y=Score_Full)) + 
  geom_jitter(alpha=0.8, shape=16, position=position_jitter(0.1), size=point.sz, aes(colour=Treatment))+
  geom_boxplot(alpha=0.5, shape=2, outlier.shape = NA, linewidth=bar.sz, aes(fill=Treatment))+
  scale_colour_manual(values =HC.colors.o)+
  scale_fill_manual(values =HC.colors.o)+
  theme_classic()+
  labs(x="Treatment", y="Color Score")+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"), 
        legend.position="none")+
  geom_rect(aes(xmin=.5, xmax=1.5, ymin=16, ymax=17.5), alpha=.01, fill="grey", colour="black", linewidth=bar.sz-.5)+
  annotate("text", x=1, y=17, label= "Color Score = ", size=sig.sz, fontface="italic")+
  annotate("text", x=1, y=16.5, label= "-1(LD1 Score) + 20", size=sig.sz, fontface="italic");Score.boxplot_fig1
```

#### Figure 1DE

```{r}
##Create Panel
Color_Score_Method_fig<-plot_grid(Color.LDA.PCA_LD1.plot_fig1, Score.boxplot_fig1,
                     nrow=1, ncol=2, 
                     rel_widths=1, rel_heights=1, 
                     labels=c("d", "e"),
                     label_size=panel.lab.sz-1,
                     label_fontface = "bold")

##Save Figure
ggsave(filename="Figures/Figure1de_Color_Score_Method.png", plot=Color_Score_Method_fig, dpi=300, width=10, height=5, units="in")

```



### Figure S2 Color Score LDAs

```{r}
##Update legends
Color.LDA.PCA_LD1.plot<-Color.LDA.PCA_LD1.plot+theme(legend.position=c(.8, .8))
Color_W2.LDA.PCA_LD1.plot<-Color_W2.LDA.PCA_LD1.plot+theme(legend.position="none")
Color_M1.LDA.PCA_LD1.plot<-Color_M1.LDA.PCA_LD1.plot+theme(legend.position="none")
Color_M4.LDA.PCA_LD1.plot<-Color_M4.LDA.PCA_LD1.plot+theme(legend.position="none")


##Create Panel
Color_Score_LDA_fig<-plot_grid(Color.LDA.PCA_LD1.plot, Color_W2.LDA.PCA_LD1.plot,
                     Color_M1.LDA.PCA_LD1.plot, Color_M4.LDA.PCA_LD1.plot, 
                     nrow=2, ncol=2, 
                     rel_widths=1, rel_heights=1, 
                     labels=c("a", "b", "c", "d"),
                     label_size=panel.lab.sz-2,
                     label_fontface = "bold")

##Save Figure
ggsave(filename="Figures/FigureS2_Color_Score_LDA.png", plot=Color_Score_LDA_fig, dpi=300, width=10, height=8, units="in")
 
```
