==> 04_Models.Rmd <==
---
title: "Relationship between Color Score and Bleaching Metrics"
author: "Serena Hackerott and Lauren Gregory"
date: "7/19/2024"
output:
  html_document:
    toc: yes
    df_print: paged
  html_notebook:
    toc: yes
    toc_float: yes
---

# Setup
```{r Setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```


### Load Packages
```{r}
##Install Packages if Needed
if (!require("ggplot2")) install.packages("ggplot2")
if (!require("effectsize")) install.packages("effectsize")
if (!require("dplyr")) install.packages("dplyr")
if (!require("ggeffects")) install.packages("ggeffects")
if (!require("cowplot")) install.packages("cowplot")
if (!require("Hmisc")) install.packages("Hmisc")

##Load Packages
library(ggplot2) #Required for plotting
library(effectsize) #Required to eta_squared effect sizes
library(dplyr) #Required for dataframe cleanup
library(ggeffects) #Required for ggpredict to plot model results
library(cowplot) #Required for arranging ggplots 
library(Hmisc) #Required for correlation matrix with p values
```
Note: Run "Graphing Parameters" section from 01_ExperimentalSetup.R file


### Load and Organize Data
Note: BleachData was created in 02_BleachingMetrics.Rmd and ScoreData was created in 03_ColorScore.Rmd file
```{r}
#Load Data
ScoreData<-read.csv("Outputs/ScoreData.csv", header=TRUE)
BleachData<-read.csv("Outputs/BleachData.csv", header=TRUE)

##Combine Bleaching and Color Score Data
FullData<-merge(ScoreData, BleachData, all.x=TRUE)

##Set factor variables 
FullData$TimeP<-factor(FullData$TimeP, levels=c("W2", "M1", "M4"), ordered=TRUE)
FullData$Site<-factor(FullData$Site, levels=c("KL", "SS"), ordered=TRUE)
FullData$Genotype<-factor(FullData$Genotype, levels=c("AC8", "AC10", "AC12"), ordered=TRUE)
FullData$Treatment<-factor(FullData$Treatment, levels=c("Control", "Heat"), ordered=TRUE)
FullData$Treat<-factor(FullData$Treat, levels=c("C", "H"), ordered=TRUE)
FullData$Seas<-factor(FullData$Seas, levels=c("Summer1", "Summer2", "Winter"), ordered=TRUE)

```


# Model Development
Biologically, traditional bleaching metrics (chlorophyll, symbiont density) should act as a predictor and color score would be the response. However to test the hypothesis if color score can be used to predict traditional bleaching metrics and to make this model more accessible to coral restoration facilities, color score will be modeled as the predictor (independent, x variable) with traditional metrics as the response (dependent, y variable).


### Checking Normality

```{r}
##Chlorophyll
hist(FullData$Chl_ug.cm2)
shapiro.test(FullData$Chl_ug.cm2)
# Not Normal

#Natural log+1 transformation 
hist(log(FullData$Chl_ug.cm2+1))
shapiro.test(log(FullData$Chl_ug.cm2+1))
# Not Normal

#Square root transformation 
hist(sqrt(FullData$Chl_ug.cm2))
shapiro.test(sqrt(FullData$Chl_ug.cm2))
# Not Normal

```


```{r}
##Symbionts
hist(FullData$Sym10.6_cm2)
shapiro.test(FullData$Sym10.6_cm2)
# Not Normal

#Natural log+1 transformation 
hist(log(FullData$Sym10.6_cm2+1))
shapiro.test(log(FullData$Sym10.6_cm2+1))
# Not Normal

#Square root transformation 
hist(sqrt(FullData$Sym10.6_cm2))
shapiro.test(sqrt(FullData$Sym10.6_cm2))
# Not Normal

```


```{r}
##Color Score with Full Dataset
hist(FullData$Score_Full)
shapiro.test(FullData$Score_Full)
# Not Normal

#Square transformation 
hist((FullData$Score_Full)^2)
shapiro.test((FullData$Score_Full)^2)
# Not Normal

```


```{r}
##Color Score by Timepoints
hist(FullData$Score_TP)
shapiro.test(FullData$Score_TP)
# Not Normal

#Square transformation 
hist((FullData$Score_TP)^2)
shapiro.test((FullData$Score_TP)^2)
# Not Normal

```

None of the variables are normally distributed. 


### Checking Correlation
Checking correlation between each Color Score calculation with traditional bleaching metrics. Using Pearson correlation to check for linear correlation. 

```{r}
names(FullData)
rcorr(as.matrix(FullData[,c("Score_Full", "Score_TP", "Chl_ug.cm2", "Sym10.6_cm2")], type="pearson"))
```

Both metrics of Color Score (calculated from different subset strategies) are highly correlated with both Chlorophyll concentration and Symbiont density. 

The two metrics of Color Score are highly correlated with each other (0.90).

Color Score calculated from the Full set and Color Score calculated from each Timepoint are similar (.74) in their correlations with Chlorophyll concentration (0.74) and with Symbiont density (0.60-0.63). Additionally, these are on par with the correlation between Chlorophyll and Symbionts (0.60)


# Model Chlorophyll ~ Color

## Full Set

### Model Selection
```{r}
#Compare Model Distributions
Chl_Full.glm.gaus<-glm(Chl_ug.cm2~Score_Full, data=FullData, family=gaussian(link="identity"))
Chl_Full.glm.gaus.l<-glm(Chl_ug.cm2~Score_Full, data=FullData, family=gaussian(link="log"))
Chl_Full.glm.gaus.i<-glm(Chl_ug.cm2~Score_Full, data=FullData, family=gaussian(link="inverse"))

Chl_Full.glm.gam.l<-glm(Chl_ug.cm2~Score_Full, data=FullData, family=Gamma(link="log"))
Chl_Full.glm.gam.i<-glm(Chl_ug.cm2~Score_Full, data=FullData, family=Gamma(link="inverse"))

#AIC Model Comparison
AIC(Chl_Full.glm.gaus, Chl_Full.glm.gaus.l, Chl_Full.glm.gaus.i, Chl_Full.glm.gam.l, Chl_Full.glm.gam.i)
```

Gamma distribution with log link has the best model fit (lowest AIC)


### Model Results
```{r}
#Model Results
summary(Chl_Full.glm.gam.l)

#Effect Size of Color Score
eta_squared(Chl_Full.glm.gam.l, partial=FALSE)

##Save model results
Chl_FUll.glm.res<-data.frame(summary(Chl_Full.glm.gam.l)$coefficients)
rownames(Chl_FUll.glm.res)[1]<-"Intercept"
Chl_FUll.glm.res$Predictor<-rownames(Chl_FUll.glm.res)
Chl_FUll.glm.res$EtaSq<-c(NA,eta_squared(Chl_Full.glm.gam.l, partial=FALSE)$Eta2)
Chl_FUll.glm.res$Response<-rep("Chlorophyll", nrow(Chl_FUll.glm.res))
Chl_FUll.glm.res<-Chl_FUll.glm.res %>% dplyr::rename( p.value = "Pr...t..")
Chl_FUll.glm.res<-Chl_FUll.glm.res %>% dplyr::rename( Std.Error = "Std..Error")
Chl_FUll.glm.res$Null.Dev<-c(summary(Chl_Full.glm.gam.l)$null.deviance, summary(Chl_Full.glm.gam.l)$df.null)
Chl_FUll.glm.res$Resid.Dev<-c(summary(Chl_Full.glm.gam.l)$deviance, summary(Chl_Full.glm.gam.l)$df.residual)

```

Color Score calculated for the Full dataset significantly predicts Chlorophyll (p< 2e-16) and explains 55% (Eta squared 0.55) of the variance. 


#### Check Residuals
```{r}
##Check Normality of Residuals
#Distribution 
plot(density(resid(Chl_Full.glm.gam.l)))

#Q-Q plot
qqnorm(resid(Chl_Full.glm.gam.l)); qqline(resid(Chl_Full.glm.gam.l))

##Check Variance of Residuals across Fitted Values
plot(fitted(Chl_Full.glm.gam.l), resid(Chl_Full.glm.gam.l))

```

Residuals look good.


### Plot Model
```{r Model of Chl vs Color Full Set}
Chl_Full.mod.plot<-plot(ggpredict(Chl_Full.glm.gam.l , terms="Score_Full"), rawdata=TRUE, colors="darkgreen", jitter=TRUE)+
  labs(x="Color Score Full Set", y=expression(paste('Chlorophyll (\u03BCg cm'^-2*") ")), title=NULL)+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"))+
        annotate("text", x=18, y=-.4, label=expression(bolditalic(paste("p < 0.001  ", eta^2, " = 0.55 "))),size=sig.sz, hjust = 0);Chl_Full.mod.plot


```


## Each Timepoint

### Model Selection
```{r}
#Compare Model Distributions
Chl_TP.glm.gaus<-glm(Chl_ug.cm2~Score_TP, data=FullData, family=gaussian(link="identity"))
Chl_TP.glm.gaus.l<-glm(Chl_ug.cm2~Score_TP, data=FullData, family=gaussian(link="log"))
Chl_TP.glm.gaus.i<-glm(Chl_ug.cm2~Score_TP, data=FullData, family=gaussian(link="inverse"))

Chl_TP.glm.gam.l<-glm(Chl_ug.cm2~Score_TP, data=FullData, family=Gamma(link="log"))
Chl_TP.glm.gam.i<-glm(Chl_ug.cm2~Score_TP, data=FullData, family=Gamma(link="inverse"))

#AIC Model Comparison
AIC(Chl_TP.glm.gaus, Chl_TP.glm.gaus.l, Chl_TP.glm.gaus.i, Chl_TP.glm.gam.l, Chl_TP.glm.gam.i)
```

Gamma distribution with log link has the best model fit (lowest AIC)


### Model Results
```{r}
#Model Results
summary(Chl_TP.glm.gam.l)

#Effect Size of Color Score
eta_squared(Chl_TP.glm.gam.l, partial=FALSE)

##Save model results
Chl_TP.glm.res<-data.frame(summary(Chl_TP.glm.gam.l)$coefficients)
rownames(Chl_TP.glm.res)[1]<-"Intercept"
Chl_TP.glm.res$Predictor<-rownames(Chl_TP.glm.res)
Chl_TP.glm.res$EtaSq<-c(NA,eta_squared(Chl_TP.glm.gam.l, partial=FALSE)$Eta2)
Chl_TP.glm.res$Response<-rep("Chlorophyll", nrow(Chl_TP.glm.res))
Chl_TP.glm.res<-Chl_TP.glm.res %>% dplyr::rename( p.value = "Pr...t..")
Chl_TP.glm.res<-Chl_TP.glm.res %>% dplyr::rename( Std.Error = "Std..Error")
Chl_TP.glm.res$Null.Dev<-c(summary(Chl_TP.glm.gam.l)$null.deviance, summary(Chl_TP.glm.gam.l)$df.null)
Chl_TP.glm.res$Resid.Dev<-c(summary(Chl_TP.glm.gam.l)$deviance, summary(Chl_TP.glm.gam.l)$df.residual)

```

Color Score calculated for the Full dataset significantly predicts Chlorophyll (p< 2e-16) and explains 54% (Eta squared 0.54) of the variance. 


#### Check Residuals
```{r}
##Check Normality of Residuals
#Distribution 
plot(density(resid(Chl_TP.glm.gam.l)))

#Q-Q plot
qqnorm(resid(Chl_TP.glm.gam.l)); qqline(resid(Chl_TP.glm.gam.l))

##Check Variance of Residuals across Fitted Values
plot(fitted(Chl_TP.glm.gam.l), resid(Chl_TP.glm.gam.l))

```

Residuals look good.


### Plot Model

```{r Model of Chl vs Color by Timepoint}
Chl_TP.mod.plot<-plot(ggpredict(Chl_TP.glm.gam.l , terms="Score_TP"), rawdata=TRUE, colors="lightgreen", jitter=TRUE)+
  labs(x="Color Score by Timepoint", y=expression(paste('Chlorophyll (\u03BCg cm'^-2*") ")), title=NULL)+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"))+
        annotate("text", x=20, y=-.4, label=expression(bolditalic(paste("p < 0.001  ", eta^2, " = 0.54 "))),size=sig.sz, hjust = 0);Chl_TP.mod.plot

```


# Model Symbionts ~ Color

## Full Set

### Model Selection
```{r}
#Compare Model Distributions
Sym_Full.glm.gaus<-glm(Sym10.6_cm2~Score_Full, data=FullData, family=gaussian(link="identity"))
Sym_Full.glm.gaus.l<-glm(Sym10.6_cm2~Score_Full, data=FullData, family=gaussian(link="log"))
Sym_Full.glm.gaus.i<-glm(Sym10.6_cm2~Score_Full, data=FullData, family=gaussian(link="inverse"))

Sym_Full.glm.gam.l<-glm(Sym10.6_cm2~Score_Full, data=FullData, family=Gamma(link="log"))
Sym_Full.glm.gam.i<-glm(Sym10.6_cm2~Score_Full, data=FullData, family=Gamma(link="inverse"))


#AIC Model Comparison
AIC(Sym_Full.glm.gaus, Sym_Full.glm.gaus.l, Sym_Full.glm.gaus.i, Sym_Full.glm.gam.l, Sym_Full.glm.gam.i)
```

Gamma distribution with log link has the best model fit (lowest AIC).


### Model Results
```{r}
#Model Results
summary(Sym_Full.glm.gam.l)

#Effect Size of Color Score
eta_squared(Sym_Full.glm.gam.l, partial=FALSE)

##Save model results
Sym_Full.glm.res<-data.frame(summary(Sym_Full.glm.gam.l)$coefficients)
rownames(Sym_Full.glm.res)[1]<-"Intercept"
Sym_Full.glm.res$Predictor<-rownames(Sym_Full.glm.res)
Sym_Full.glm.res$EtaSq<-c(NA,eta_squared(Sym_Full.glm.gam.l, partial=FALSE)$Eta2)
Sym_Full.glm.res$Response<-rep("Symbionts", nrow(Sym_Full.glm.res))
Sym_Full.glm.res<-Sym_Full.glm.res %>% dplyr::rename( p.value = "Pr...t..")
Sym_Full.glm.res<-Sym_Full.glm.res %>% dplyr::rename( Std.Error = "Std..Error")
Sym_Full.glm.res$Null.Dev<-c(summary(Sym_Full.glm.gam.l)$null.deviance, summary(Sym_Full.glm.gam.l)$df.null)
Sym_Full.glm.res$Resid.Dev<-c(summary(Sym_Full.glm.gam.l)$deviance, summary(Sym_Full.glm.gam.l)$df.residual)

```

Color Score calculated for the Full dataset significantly predicts Chlorophyll (p< 2e-16) and explains 40% (Eta squared 0.40) of the variance. 


#### Check Residuals
```{r}
##Check Normality of Residuals
#Distribution 
plot(density(resid(Sym_Full.glm.gam.l)))

#Q-Q plot
qqnorm(resid(Sym_Full.glm.gam.l)); qqline(resid(Sym_Full.glm.gam.l))

##Check Variance of Residuals across Fitted Values
plot(fitted(Sym_Full.glm.gam.l), resid(Sym_Full.glm.gam.l))

```

Residuals look good.


### Plot Model
```{r Model of Sym vs Color Full Set}
Sym_Full.mod.plot<-plot(ggpredict(Sym_Full.glm.gam.l , terms="Score_Full"), rawdata=TRUE, colors="darkorange3", jitter=TRUE)+
  labs(x="Color Score Full Set", y=expression(paste('Symbionts (*10'^-6* ' cm'^-2*") ")), title=NULL)+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"))+
        annotate("text", x=18, y=-.2, label=expression(bolditalic(paste("p < 0.001  ", eta^2, " = 0.40 "))),size=sig.sz, hjust = 0);Sym_Full.mod.plot


```


## Each Timepoint

### Model Selection
```{r}
#Compare Model Distributions
Sym_TP.glm.gaus<-glm(Sym10.6_cm2~Score_TP, data=FullData, family=gaussian(link="identity"))
Sym_TP.glm.gaus.l<-glm(Sym10.6_cm2~Score_TP, data=FullData, family=gaussian(link="log"))
Sym_TP.glm.gaus.i<-glm(Sym10.6_cm2~Score_TP, data=FullData, family=gaussian(link="inverse"))

Sym_TP.glm.gam.l<-glm(Sym10.6_cm2~Score_TP, data=FullData, family=Gamma(link="log"))
Sym_TP.glm.gam.i<-glm(Sym10.6_cm2~Score_TP, data=FullData, family=Gamma(link="inverse"))

#AIC Model Comparison
AIC(Sym_TP.glm.gaus, Sym_TP.glm.gaus.l, Sym_TP.glm.gaus.i, Sym_TP.glm.gam.l, Sym_TP.glm.gam.i)
```

Gamma distribution with log link has the best model fit (lowest AIC).


### Model Results
```{r}
#Model Results
summary(Sym_TP.glm.gam.l)

#Effect Size of Color Score
eta_squared(Sym_TP.glm.gam.l, partial=FALSE)

##Save model results
Sym_TP.glm.res<-data.frame(summary(Sym_TP.glm.gam.l)$coefficients)
rownames(Sym_TP.glm.res)[1]<-"Intercept"
Sym_TP.glm.res$Predictor<-rownames(Sym_TP.glm.res)
Sym_TP.glm.res$EtaSq<-c(NA,eta_squared(Sym_TP.glm.gam.l, partial=FALSE)$Eta2)
Sym_TP.glm.res$Response<-rep("Symbionts", nrow(Sym_TP.glm.res))
Sym_TP.glm.res<-Sym_TP.glm.res %>% dplyr::rename( p.value = "Pr...t..")
Sym_TP.glm.res<-Sym_TP.glm.res %>% dplyr::rename( Std.Error = "Std..Error")
Sym_TP.glm.res$Null.Dev<-c(summary(Sym_TP.glm.gam.l)$null.deviance, summary(Sym_TP.glm.gam.l)$df.null)
Sym_TP.glm.res$Resid.Dev<-c(summary(Sym_TP.glm.gam.l)$deviance, summary(Sym_TP.glm.gam.l)$df.residual)

```

Color Score calculated for the Full dataset significantly predicts Chlorophyll (p< 2e-16) and explains 37% (Eta squared 0.37) of the variance. 


#### Check Residuals
```{r}
##Check Normality of Residuals
#Distribution 
plot(density(resid(Sym_TP.glm.gam.l)))

#Q-Q plot
qqnorm(resid(Sym_TP.glm.gam.l)); qqline(resid(Sym_TP.glm.gam.l))

##Check Variance of Residuals across Fitted Values
plot(fitted(Sym_TP.glm.gam.l), resid(Sym_TP.glm.gam.l))

```

Residuals look good.


### Plot Model

```{r Model of Sym vs Color by Timepoint}
Sym_TP.mod.plot<-plot(ggpredict(Sym_TP.glm.gam.l , terms="Score_TP"), rawdata=TRUE, colors="darkorange", jitter=TRUE)+
  labs(x="Color Score by Timepoint", y=expression(paste('Symbionts (*10'^-6* ' cm'^-2*") ")), title=NULL)+
  theme_classic()+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"))+
        annotate("text", x=20, y=-.2, label=expression(bolditalic(paste("p < 0.001  ", eta^2, " = 0.37 "))),size=sig.sz, hjust = 0);Sym_TP.mod.plot

```


# Model Predictions
Using similar models of Bleaching Metric ~ Color Score developed for Summer 1 (Timepoint W2) to test the ability to use Color Score to predict the Bleaching Metric in Summer 2 (Timepoint M1) and Winter (Timepoint M4) seasonal timepoints.

## Model Chl ~ Color 

### Develop Summer 1 Model
```{r}
##Model with Gamma Log Link
Chl_W2.glm<-glm(Chl_ug.cm2~Score_TP, data=subset(FullData, TimeP=="W2"), family=Gamma(link="log"))


##Check Residuals
#Distribution 
plot(density(resid(Chl_W2.glm)))

#Q-Q plot
qqnorm(resid(Chl_W2.glm)); qqline(resid(Chl_W2.glm))

##Residuals vs Fitted Values
plot(fitted(Chl_W2.glm), resid(Chl_W2.glm))

##Model Results
summary(Chl_W2.glm)

#Effect Size of Color Score
eta_squared(Chl_W2.glm, partial=FALSE)

##Save model results
Chl_W2.glm.res<-data.frame(summary(Chl_W2.glm)$coefficients)
Chl_W2.glm.res$Predictor<-c("Intercept", "Color Summer 1")
Chl_W2.glm.res$EtaSq<-c(NA,eta_squared(Chl_W2.glm, partial=FALSE)$Eta2)
Chl_W2.glm.res$Response<-rep("Chlorophyll Summer 1", nrow(Chl_W2.glm.res))
Chl_W2.glm.res<-Chl_W2.glm.res %>% dplyr::rename( p.value = "Pr...t..")
Chl_W2.glm.res<-Chl_W2.glm.res %>% dplyr::rename( Std.Error = "Std..Error")
Chl_W2.glm.res$Null.Dev<-c(summary(Chl_W2.glm)$null.deviance, summary(Chl_W2.glm)$df.null)
Chl_W2.glm.res$Resid.Dev<-c(summary(Chl_W2.glm)$deviance, summary(Chl_W2.glm)$df.residual)

```


### Predict Summer 2
```{r}
##Predicted values on Response scale 
M1_Chl_predict<-data.frame(Predicted=predict(Chl_W2.glm, newdata=data.frame(Score_TP=c(FullData$Score_TP[which(FullData$TimeP=="M1")])), type="response"))

##Compare with Measured values 
M1_Chl_predict$Measured<-FullData$Chl_ug.cm2[which(FullData$TimeP=="M1")]

##Test correlation between Predicted and Measured values
cor.test(M1_Chl_predict$Predicted, M1_Chl_predict$Measured, method="pearson")

```

The measured values of Summer 2 Chlorophyll are significantly (p < 0.001) correlated with those predicted from Color Score using the model developed in Summer 1.


#### Plot Predicted vs Measured
```{r}
Chl_M1.pred.plot<-ggplot(M1_Chl_predict, aes(x=Predicted, y=Measured))+
    geom_point(size=point.sz, colour="lightgreen", pch=15)+ 
  labs(x=expression(paste('Predicted Chlorophyll (\u03BCg cm'^-2*") ")), y="Measured Chlorophyll", title="Summer")+
  theme_classic()+
  ylim(0,2.4)+
  xlim(0,2.4)+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        plot.title = element_text(size = plot.title.sz, colour="black", hjust = 0.5))+
  annotate("text",  x=1.3, y=0, label="r = 0.886  p < 0.001", fontface="bold.italic",size=sig.sz, hjust = 0);Chl_M1.pred.plot

```


### Predict Winter
```{r}
##Predicted values on Response scale
M4_Chl_predict<-data.frame(Predicted=predict(Chl_W2.glm, newdata=data.frame(Score_TP=c(FullData$Score_TP[which(FullData$TimeP=="M4")])), type="response"))

##Compare with Measured values 
M4_Chl_predict$Measured<-FullData$Chl_ug.cm2[which(FullData$TimeP=="M4")]

##Test correlation between Predicted and Measured values
cor.test(M4_Chl_predict$Predicted, M4_Chl_predict$Measured, method="pearson")

```

The measured values of Winter Chlorophyll are significantly (p < 0.001) correlated with those predicted from Color Score using the model developed in Summer 1.


#### Plot Predicted vs Measured
```{r}
Chl_M4.pred.plot<-ggplot(M4_Chl_predict, aes(x=Predicted, y=Measured))+
    geom_point(size=point.sz, colour="lightgreen", pch=17)+ 
  labs(x=expression(paste('Predicted Chlorophyll (\u03BCg cm'^-2*") ")), y="Measured Chlorophyll", title="Winter")+
  theme_classic()+
  ylim(0,2)+
  xlim(0,2)+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        plot.title = element_text(size = plot.title.sz, colour="black", hjust = 0.5))+
  annotate("text",  x=1, y=0, label="r = 0.680  p < 0.001", fontface="bold.italic",size=sig.sz, hjust = 0);Chl_M4.pred.plot

```


## Model Sym ~ Color 

### Develop Summer 1 Model
```{r}
##Model with Gamma Log Link
Sym_W2.glm<-glm(Sym10.6_cm2~Score_TP, data=subset(FullData, TimeP=="W2"), family=Gamma(link="log"))


##Check Residuals
#Distribution 
plot(density(resid(Sym_W2.glm)))

#Q-Q plot
qqnorm(resid(Sym_W2.glm)); qqline(resid(Sym_W2.glm))

##Residuals vs Fitted Values
plot(fitted(Sym_W2.glm), resid(Sym_W2.glm))

##Model Results
summary(Sym_W2.glm)

#Effect Size of Color Score
eta_squared(Sym_W2.glm, partial=FALSE)

##Save model results
Sym_W2.glm.res<-data.frame(summary(Sym_W2.glm)$coefficients)
Sym_W2.glm.res$Predictor<-c("Intercept", "Color Summer 1")
Sym_W2.glm.res$EtaSq<-c(NA,eta_squared(Sym_W2.glm, partial=FALSE)$Eta2)
Sym_W2.glm.res$Response<-rep("Symbionts Summer 1", nrow(Sym_W2.glm.res))
Sym_W2.glm.res<-Sym_W2.glm.res %>% dplyr::rename( p.value = "Pr...t..")
Sym_W2.glm.res<-Sym_W2.glm.res %>% dplyr::rename( Std.Error = "Std..Error")
Sym_W2.glm.res$Null.Dev<-c(summary(Sym_W2.glm)$null.deviance, summary(Sym_W2.glm)$df.null)
Sym_W2.glm.res$Resid.Dev<-c(summary(Sym_W2.glm)$deviance, summary(Sym_W2.glm)$df.residual)

```


### Predict Summer 2
```{r}
##Predicted values on Response scale
M1_Sym_predict<-data.frame(Predicted=predict(Sym_W2.glm, newdata=data.frame(Score_TP=c(FullData$Score_TP[which(FullData$TimeP=="M1")])), type="response"))

##Compare with Measured values 
M1_Sym_predict$Measured<-FullData$Sym10.6_cm2[which(FullData$TimeP=="M1")]

##Test correlation between Predicted and Measured values
cor.test(M1_Sym_predict$Predicted, M1_Sym_predict$Measured, method="pearson")

```

The measured values of Summer 2 Symbiont Density are significantly (p < 0.001) correlated with those predicted from Color Score using the model developed in Summer 1.


#### Plot Predicted vs Measured
```{r}
Sym_M1.pred.plot<-ggplot(M1_Sym_predict, aes(x=Predicted, y=Measured))+
    geom_point(size=point.sz, colour="darkorange", pch=15)+ 
  labs(x=expression(paste('Predicted Symbionts (*10'^-6* ' cm'^-2*") ")), y="Measured Symbionts", title="Summer")+
  theme_classic()+
  xlim(0, 1.3)+
  ylim(0, 1.3)+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        plot.title = element_text(size = plot.title.sz, colour="black", hjust = 0.5))+
   annotate("text",  x=0.6, y=0, label="r = 0.754  p < 0.001", fontface="bold.italic",size=sig.sz, hjust = 0);Sym_M1.pred.plot

```


### Predict Winter
```{r}
##Predicted values on Response scale 
M4_Sym_predict<-data.frame(Predicted=predict(Sym_W2.glm, newdata=data.frame(Score_TP=c(FullData$Score_TP[which(FullData$TimeP=="M4")])), type="response"))

##Compare with Measured values 
M4_Sym_predict$Measured<-FullData$Sym10.6_cm2[which(FullData$TimeP=="M4")]

##Test correlation between Predicted and Measured values
cor.test(M4_Sym_predict$Predicted, M4_Sym_predict$Measured, method="pearson")

```

The measured values of Winter Symbiont Density are significantly (p < 0.05) correlated with those predicted from Color Score using the model developed in Summer 1.


#### Plot Predicted vs Measured
```{r}
Sym_M4.pred.plot<-ggplot(M4_Sym_predict, aes(x=Predicted, y=Measured))+
    geom_point(size=point.sz, colour="darkorange", pch=17)+ 
  labs(x=expression(paste('Predicted Symbionts (*10'^-6* ' cm'^-2*") ")), y="Measured Symbionts", title="Winter")+
  theme_classic()+
 xlim(0, 1)+
  ylim(0, 1)+
  theme(axis.title.x = element_text(size = axis.title.sz), 
        axis.title.y = element_text(size = axis.title.sz), 
        axis.text.x=element_text(size=axis.txt.sz, colour="black"),
        axis.text.y=element_text(size=axis.txt.sz, colour="black"),
        plot.title = element_text(size = plot.title.sz, colour="black", hjust = 0.5))+
   annotate("text",  x=0.5, y=0, label="r = 0.345  p < 0.05", fontface="bold.italic",size=sig.sz, hjust = 0);Sym_M4.pred.plot

```


# Write Out FullData
```{r}
write.csv(FullData, "Outputs/FullData.csv", row.names=FALSE)

```



# Figures

### Figure 2 Bleaching vs Color Score
```{r}
##Create Panel
Bleach_Color_Models_fig<-plot_grid(Chl_Full.mod.plot, Chl_TP.mod.plot,
                     Sym_Full.mod.plot, Sym_TP.mod.plot, 
                     nrow=2, ncol=2, 
                     rel_widths=1, rel_heights=1, 
                     labels=c("a", "b", "c", "d"),
                     label_size=panel.lab.sz-2,
                     label_fontface = "bold")

##Save Figure
ggsave(filename="Figures/Figure2_Bleaching_Color_Models.png", plot=Bleach_Color_Models_fig, dpi=300, width=10, height=8, units="in")
 
```


### Figure 3 Predicted vs Measured Bleaching

```{r}
##Remove Titles for Panel
Sym_M1.pred.plot<-Sym_M1.pred.plot+labs(title=NULL)
Sym_M4.pred.plot<-Sym_M4.pred.plot+labs(title=NULL)

##Create Panel
Bleach_Predict_fig<-plot_grid(Chl_M1.pred.plot, Chl_M4.pred.plot,
                     Sym_M1.pred.plot, Sym_M4.pred.plot, 
                     nrow=2, ncol=2, 
                     rel_widths=1, rel_heights=1, 
                     labels=c("a", "b", "c", "d"),
                     label_size=panel.lab.sz-2,
                     label_fontface = "bold")

##Save Figure
ggsave(filename="Figures/Figure3_Bleaching_Predicted_Measured.png", plot=Bleach_Predict_fig, dpi=300, width=10, height=8, units="in")
 
```


# Tables

### Table S1
```{r}
##Combine Results Tables
TableS1_glm.res<-rbind(Chl_FUll.glm.res, Chl_TP.glm.res, Sym_Full.glm.res, Sym_TP.glm.res, 
                       Chl_W2.glm.res, Sym_W2.glm.res)

##Organize
names(TableS1_glm.res)
TableS1_Mod.res<-TableS1_glm.res[,c("Response", "Null.Dev" , "Resid.Dev", "Predictor", "Estimate", "Std.Error", "t.value",   "p.value", "EtaSq")]
TableS1_Mod.res<-TableS1_Mod.res %>% dplyr::rename( SE = Std.Error)

#Round to 4 digits
TableS1_Mod.res$Null.Dev<-round(TableS1_Mod.res$Null.Dev, 4)
TableS1_Mod.res$Resid.Dev<-round(TableS1_Mod.res$Resid.Dev, 4)
TableS1_Mod.res$SE<-round(TableS1_Mod.res$SE, 4)
TableS1_Mod.res$t.value<-round(TableS1_Mod.res$t.value, 4)
TableS1_Mod.res$EtaSq<-round(TableS1_Mod.res$EtaSq, 4)
TableS1_Mod.res$p.value<-round(TableS1_Mod.res$p.value, 4)

##Write Out Table
write.csv(TableS1_Mod.res, "Tables/TableS1_Metric_Comparison_GLM_Results.csv", row.names=FALSE)

```
